+++ CHANGELOG.md
@@ -6,10 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Change
-- Turn eslint back on / fix Vue SFC CSS character error when eslint enabled
-- Use newer @rollup/babel
 - Move fetch posts into a services file
 - Add tests
+
+## [0.0.3]
+### Changed
+- ESLint works now
+- Newer @rollup/babel added
 
 ## [0.0.2]
 ### Changed